+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Geolocation field rules not being considered on form validation.
 - Geolocation number input not being required (should have a value or the not applicable value).
 
+### Added
+
+- `useGeolocation` prop for `AddressRules` component.
+
 ## [3.6.13] - 2019-11-06
 
 ### Changed